Updated all tests for action creators. Also moved some things into restClient and restClient.test.js
