+++ src/common/utils/restClient.test.js
@@ -1,5 +1,21 @@
 import { fetchJsonData } from '../restClient';
 
+// import urls from '../../urls';
+
+// import * as types from '../types';
+
+// global.fetch = jest.fn(() =>
+//   setTimeout( ()=>{
+//     Promise.resolve({
+//       json: ()=> Promise.resolve({ data: { main: { temp: 50 }}}),
+//     })
+//   }, 100)
+// );
+//
+// test('and calls the correct URL.', ()=>{
+//   expect(global.fetch).toHaveBeenCalledWith(urls.openWeatherUrl);
+// });
+
 describe('restClient.js / fetchJsonData()', ()=>{
   describe('WHEN: given an endpoint, ', ()=>{
     it('THEN: fetches the JSON data.', async ()=>{
