+++ src/reducers/globalHumidityReducer.js
@@ -1,4 +1,4 @@
-import { FETCH_HUMIDITY } from '../actions/types';
+import { FETCH_HUMIDITY } from '../actionCreators/types';
 
 export default function globalHumidityReducer(state = 0, action){
   switch (action.type){
