+++ src/reducers/globalTempReducer.js
@@ -1,4 +1,4 @@
-import { FETCH_TEMP } from '../actions/types';
+import { FETCH_TEMP } from '../actionCreators/types';
 
 export default function globalTempReducer(state = 0, action){
   switch (action.type){
